requirements.txt
fastapi
uvicorn
twilio
langgraph
langchain
langchain-groq
langchain-core
langchain-community
pydantic
python-dotenv
tavily-python
python-docx
